acp-ingest/requirements.txt
# Development tools
pre-commit==3.6.0

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
gunicorn==21.2.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
cryptography>=42.0.0
bcrypt==4.1.2

# Database and storage
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
chromadb==0.4.18

# Vault integration for secrets management
hvac==2.0.0

# Security: Safe XML parsing
defusedxml==0.7.1

# File processing and parsing
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1

# OCR and image processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# PII detection and redaction
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.2
transformers==4.36.2

# HTTP client and utilities
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.0

# CLI interface
click==8.1.7
typer==0.9.0
rich==13.7.0

# Background jobs and async
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# OpenTelemetry and tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-celery==0.42b0

# Vault integration
hvac==2.0.0
python-json-logger==2.0.7
psutil==5.9.6

# Export and reporting
jinja2==3.1.2
reportlab==4.0.7
xlsxwriter==3.1.9

# Rate limiting and CORS
slowapi==0.1.9

# Additional security and utilities
python-magic==0.4.27
chardet==5.2.0
email-validator==2.1.0
python-slugify==8.0.1

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
factory-boy==3.3.0

# Error tracking and APM
sentry-sdk[fastapi]==1.38.0

# Text processing and NLP
nltk==3.8.1
textstat==0.7.3
langdetect==1.0.9

# Machine learning for enhanced PII detection
scikit-learn==1.3.2

# JSON utilities
orjson==3.9.10

# Time utilities
arrow==1.3.0

# UUID utilities
shortuuid==1.0.11

# Caching
cachetools==5.3.2

# Serialization
msgpack==1.0.7

# Compression
zstandard==0.22.0

# Network utilities
dnspython==2.4.2
